refactor(core): 検索スコアリングと順位計算を search/scoring.rs へ分割 (#600)
マッチ方式・bitmask pre-filter・履歴加点・tie-break・BinaryHeap 順序・所有
SearchResult 変換という「検索順位」の責務を、オーケストレーション/構築から分離する。
製品挙動・スコア定数・tie-break・並列方式は不変。

- 移設: `mod score_tier`（+ `const _` 全順序アサーション）/ thread-local `MATCHER` /
  `EntryView` + `entry_view` / `score_one_entry` / `ScoredEntry` + `Ord` 一式 /
  `heap_into_results` / `adjusted_history_boost` / `kana_substring_score` /
  `match_score_single_cached`
- 候補選択・rayon fold/reduce・incremental cache 更新（prev_*）・decide_incremental は
  search.rs に残す。TopK 抽象化は行わず BinaryHeap 更新はそのまま（後続 issue）
- 子から親の並列 Vec private を直接読む。search.rs / test から参照される
  `EntryView` / `ScoredEntry`（+フィールド）/ `score_one_entry` / `entry_view` /
  `heap_into_results` / `adjusted_history_boost` を `pub(super)`（crate へは広げない）。
  `match_score_single_cached` / `kana_substring_score` / `MATCHER` / `score_tier` は private
- ranking テストは `adjusted_history_boost` を scoring パスから import
- move で不要化した search.rs の import 2 つ（`std::cell::RefCell` / nucleo_matcher 群）を除去
- snotra-core/CLAUDE.md の search.rs 索引に scoring.rs を追記

不変条件: Prefix > Substring > Kana > Path > Fuzzy / bitmask pre-filter を entry_view 前 /
has_dot 短絡 / 直接マッチ優先 / 履歴・folder boost 式 / ScoredEntry::Ord(better=Less,
peek=worst) / String clone は top-k 確定後 K 件。

検証: cargo test -p snotra-core（466 passed, 9 ignored。受け入れ 5 テスト top_k×2 /
rank_cmp_breaks_full_tie / kana_search_direct_match_ranks / path_match_score_below_name を含む）/
clippy / rustdoc(-D warnings) / cargo check -p snotra / governance G1..G9 /
同一マシン release ベンチ（Fuzzy full scan・構築とも変更前後で回帰なし・after≤before）。
